+++ packages/datastore/src/sync/processors/sync.ts
@@ -227,6 +227,7 @@ class SyncProcessor {
 					// Catch client-side (GraphQLAuthError) & 401/403 errors here so that we don't continue to retry
 					const clientOrForbiddenErrorMessage =
 						getClientSideAuthError(error) || getForbiddenError(error);
+
 					if (clientOrForbiddenErrorMessage) {
 						logger.error('Sync processor retry error:', error);
 						throw new NonRetryableError(clientOrForbiddenErrorMessage);
@@ -284,20 +285,44 @@ class SyncProcessor {
 						});
 					}
 
+					/**
+					 * Handle $util.unauthorized() in resolver request mapper, which responses with something
+					 * like this:
+					 *
+					 * ```
+					 * {
+					 * 	data: { syncYourModel: null },
+					 * 	errors: [
+					 * 		{
+					 * 			path: ['syncLegacyJSONComments'],
+					 * 			data: null,
+					 * 			errorType: 'Unauthorized',
+					 * 			errorInfo: null,
+					 * 			locations: [{ line: 2, column: 3, sourceName: null }],
+					 * 			message:
+					 * 				'Not Authorized to access syncYourModel on type Query',
+					 * 			},
+					 * 		],
+					 * 	}
+					 * ```
+					 *
+					 * The correct handling for this is to signal that we've encountered a non-retryable error,
+					 * since the server has responded with an auth error and *NO DATA* at this point.
+					 */
 					if (unauthorized) {
-						logger.warn(
-							'queryError',
-							`User is unauthorized to query ${opName}, some items could not be returned.`
-						);
-
-						result.data = result.data || {};
-
-						result.data[opName] = {
-							...opResultDefaults,
-							...result.data[opName],
-						};
-
-						return result;
+						this.errorHandler({
+							recoverySuggestion:
+								'Ensure app code is up to date, auth directives exist and are correct on each model, and that server-side data has not been invalidated by a schema change. If the problem persists, search for or create an issue: https://github.com/aws-amplify/amplify-js/issues',
+							localModel: null!,
+							message: error.message,
+							model: modelDefinition.name,
+							operation: opName,
+							errorType: getSyncErrorType(error.errors[0]),
+							process: ProcessName.sync,
+							remoteModel: null!,
+							cause: error,
+						});
+						throw new NonRetryableError(error);
 					}
 
 					if (result.data?.[opName].items?.length) {
@@ -405,7 +430,15 @@ class SyncProcessor {
 										} catch (e) {
 											logger.error('Sync error handler failed with:', e);
 										}
-										return res();
+										/**
+										 * If there's an error, this model fails, but the rest of the sync should
+										 * continue. To facilitate this, we explicitly mark this model as `done`
+										 * with no items and allow the loop to continue organically. This ensures
+										 * all callbacks (subscription messages) happen as normal, so anything
+										 * waiting on them knows the model is as done as it can be.
+										 */
+										done = true;
+										items = [];
 									}
 
 									recordsReceived += items.length;