+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.30](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.0.29...@aws-amplify/geo@2.0.30) (2023-04-27)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## [2.0.29](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@2.0.28...@aws-amplify/geo@2.0.29) (2023-04-20)
 
 **Note:** Version bump only for package @aws-amplify/geo